Add Room implementation to retrieve ToDos from SQLite
This is a bigger commit, to fix build errors that occured when using JDK 11 instead of 8. This commit also entails changes to how the ToDo repository retrieves its data (which is now from SQLite and not in-memory). Currently this commit breaks the app, since the ViewModels still use the old repo implementation.